move frconstants.js to /config/strings.js
