+++ src/js/config/strings.js
